# Server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cinema Authentication API",
    "description": "API cho hệ thống xác thực người dùng của ứng dụng Cinema",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Đăng ký tài khoản mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email của người dùng"
                  },
                  "fullname": {
                    "type": "string",
                    "description": "Họ và tên người dùng"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Mật khẩu"
                  }
                },
                "required": ["email", "fullname", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng ký thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Đăng ký thành công. Vui lòng kiểm tra email để xác thực tài khoản."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi đăng ký",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "Email đã được đăng ký"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Đăng nhập",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "example": "bearer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "Email hoặc mật khẩu không chính xác"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/verify-email": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Xác thực email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Xác thực thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Xác thực tài khoản thành công"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "Token không hợp lệ"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/forgot-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Yêu cầu đặt lại mật khẩu",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gửi email thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email đặt lại mật khẩu đã được gửi"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "Email không tồn tại"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reset-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Đặt lại mật khẩu",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "new_password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": ["token", "new_password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đặt lại mật khẩu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Đặt lại mật khẩu thành công"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "Token không hợp lệ"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/showtimes": {
      "get": {
        "tags": ["Showtimes"],
        "summary": "Get all showtimes data",
        "responses": {
          "200": {
            "description": "Complete showtime data including all days",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "today": {
                      "$ref": "#/components/schemas/DayShowtimes"
                    },
                    "tomorrow": {
                      "$ref": "#/components/schemas/DayShowtimes"
                    },
                    "wednesday": {
                      "$ref": "#/components/schemas/DayShowtimes"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No showtime data found" 
          }
        }
      }
    },
    "/api/showtimes/by-date/{date}": {
      "get": {
        "tags": ["Showtimes"],
        "summary": "Get showtimes for a specific date",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Date in format MM/DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Showtime data for the specified date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayShowtimes"
                }
              }
            }
          },
          "404": {
            "description": "No showtimes found for the specified date"
          }
        }
      }
    },
...
    "/api/seats/{showtime_id}": {
      "get": {
        "tags": ["Seats"],
        "summary": "Get seats status for a specific showtime",
        "parameters": [
          {
            "name": "showtime_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of seats and their status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "seat_id": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string",
                        "enum": ["available", "booked", "reserved"]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}